# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Translation Bot Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface Admin {
  id: number;
  username: string;
  createdAt: string;
}

export interface AdminInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface PasswordUpdate {
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  createdAt: string;
  /** @nullable */
  lastActive?: string | null;
  totalTranslations: number;
  totalMessages: number;
  isBanned: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type TranslationDirection = typeof TranslationDirection[keyof typeof TranslationDirection];


export const TranslationDirection = {
  'bn-en': 'bn-en',
  'en-bn': 'en-bn',
} as const;

export interface Translation {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  telegramUserId?: string | null;
  /** @nullable */
  username?: string | null;
  originalText: string;
  translatedText: string;
  direction: TranslationDirection;
  createdAt: string;
  success: boolean;
  /** @nullable */
  chatType?: string | null;
}

export interface TranslationListResponse {
  translations: Translation[];
  total: number;
  page: number;
  limit: number;
}

export interface StatsOverview {
  totalUsers: number;
  activeTodayUsers: number;
  totalMessages: number;
  totalTranslations: number;
  successRate: number;
  bnToEnCount: number;
  enToBnCount: number;
  bannedUsers: number;
}

export interface DayStat {
  date: string;
  messages: number;
  translations: number;
  users: number;
}

export interface WeekStat {
  week: string;
  messages: number;
  translations: number;
  users: number;
}

export interface MonthStat {
  month: string;
  messages: number;
  translations: number;
  users: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
}

export interface BotSettings {
  translationEnabled: boolean;
  loggingEnabled: boolean;
  maintenanceMode: boolean;
  botToken: string;
  rateLimitPerMinute: number;
}

export interface BotSettingsInput {
  translationEnabled?: boolean;
  loggingEnabled?: boolean;
  maintenanceMode?: boolean;
  botToken?: string;
  rateLimitPerMinute?: number;
}

export interface TelegramUpdate { [key: string]: unknown }

export interface BotStatus {
  running: boolean;
  /** @nullable */
  botUsername: string | null;
  uptime: number;
}

export type ListUsersParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
banned?: boolean | null;
};

export type ListTranslationsParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
direction?: ListTranslationsDirection;
/**
 * @nullable
 */
userId?: number | null;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
};

export type ListTranslationsDirection = typeof ListTranslationsDirection[keyof typeof ListTranslationsDirection] | null;


export const ListTranslationsDirection = {
  'bn-en': 'bn-en',
  'en-bn': 'en-bn',
} as const;

export type ExportTranslationsParams = {
/**
 * @nullable
 */
direction?: string | null;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
};