# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Translation Bot Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List admin accounts
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create a new admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete an admin account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Admin deleted
  /admins/{id}/password:
    patch:
      operationId: updateAdminPassword
      tags: [admins]
      summary: Update admin password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Bot Webhook
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    Admin:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
    AdminInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
    PasswordUpdate:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6

    # Users
...