# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Translation Bot Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  username: string;
  createdAt: string;
}